requirements.txt
tensorflow>=2.10
numpy
pandas
opencv-python
matplotlib
streamlit
pymongo
reportlab
scikit-learn
tqdm
Pillow
tensorflow-addons
